[DAD-869] refactor: scrap repository 폴더를 각각의 카테고리 폴더로 분리
