+++ src/main/java/com/forever/dadamda/repository/scrap/video/VideoRepository.java
@@ -1,4 +1,4 @@
-package com.forever.dadamda.repository.scrap;
+package com.forever.dadamda.repository.scrap.video;
 
 import com.forever.dadamda.entity.scrap.Video;
 import com.forever.dadamda.entity.user.User;
